add flag to disable rocks write sync
